Enable allWarningsAsErrors again

+++ libraries/apollo-adapters/src/jsMain/kotlin/com/apollographql/apollo3/adapter/BigDecimal.kt
@@ -70,7 +70,7 @@ actual class BigDecimal {
   }
 
   fun toChar(): Char {
-    return jsNumber(raw).toChar()
+    return jsNumber(raw).toInt().toChar()
   }
 
   fun toDouble(): Double {
